+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Functionality in DoMINO to take arbitrary number of `scalar` or `vector`
   global parameters and encode them using `class ParameterModel`
 - TopoDiff model and example.
+- Diffusion models, metrics, and utils: implementation of Student-t
+  distribution for EDM-based diffusion models (t-EDM). This feature is adapted
+  from the paper [Heavy-Tailed Diffusion Models, Pandey et al.](https://arxiv.org/abs/2410.14171>).
+  This includes a new EDM preconditioner (`tEDMPrecondSuperRes`), a loss
+  function (`tEDMResidualLoss`), and a new option in corrdiff `diffusion_step`.
+  :warning: This is an experimental feature that can be accessed through the
+  `physicsnemo.experimental` module; it might also be subjected to API changes
+  without notice.
 
 ### Changed
 
@@ -54,6 +62,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   warnings are raised when attempting to use Apex group norm on CPU.
 - Diffusion utils: systematic compilation of patching operations in `stochastic_sampler`
   for improved performance.
+- CorrDiff example: added option for Student-t EDM (t-EDM) in `train.py` and
+  `generate.py`. When training a CorrDiff diffusion model, this feature can be
+  enabled with the hydra overrides `++training.hp.distribution=student_t` and
+  `++training.hp.nu_student_t=<nu_value>`. For generation, this feature can be
+  enabled with similar overrides: `++generation.distribution=student_t` and
+  `++generation.nu_student_t=<nu_value>`.
+- CorrDiff example: the parameters `P_mean` and `P_std` (used to compute the
+  noise level `sigma`) are now configurable. They can be set with the hydra
+  overrides `++training.hp.P_mean=<P_mean_value>` and
+  `++training.hp.P_std=<P_std_value>` for training (and similar ones with
+  `training.hp` replaced by `generation` for generation).
 
 ### Deprecated